Move all jobs to root namespace
