--- app/sidekiq/plan/plan_errored_job.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module Plan
-  class PlanErroredJob
-    include Sidekiq::Job
-
-    def perform(*args); end
-  end
-end
