--- app/sidekiq/plan/plan_running_job.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module Plan
-  class PlanRunningJob
-    include Sidekiq::Job
-
-    def perform(*args); end
-  end
-end
