+++ app/sidekiq/plan_errored_job.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class PlanErroredJob
+  include Sidekiq::Job
+
+  def perform(*args); end
+end
