+++ app/sidekiq/plan_running_job.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class PlanRunningJob
+  include Sidekiq::Job
+
+  def perform(*args); end
+end
