+++ app/sidekiq/run_stage/pre_plan_running_job.rb
@@ -9,7 +9,7 @@ module RunStage
 
       @run.task_stages.each do |task_stage|
         if task_stage.stage == 'pre_plan'
-          TaskStage::ExecuteTaskStageJob.perform_async(task_stage.id)
+          ExecuteTaskStageJob.perform_async(task_stage.id)
           next
         end
       end
