--- app/sidekiq/task_result/task_result_errored_job.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module TaskResult
-  class TaskResultErroredJob
-    include Sidekiq::Job
-
-    def perform(*args); end
-  end
-end
