--- app/sidekiq/task_result/task_result_failed_job.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module TaskResult
-  class TaskResultFailedJob
-    include Sidekiq::Job
-
-    def perform(*args); end
-  end
-end
