+++ app/sidekiq/task_result_failed_job.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class TaskResultFailedJob
+  include Sidekiq::Job
+
+  def perform(*args); end
+end
