--- app/sidekiq/task_stage/task_stage_errored_job.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module TaskStage
-  class TaskStageErroredJob
-    include Sidekiq::Job
-
-    def perform(*args); end
-  end
-end
