--- app/sidekiq/task_stage/task_stage_failed_job.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module TaskStage
-  class TaskStageFailedJob
-    include Sidekiq::Job
-
-    def perform(*_args)
-      Rails.logger.debug 'Ope, our task stage failed'
-    end
-  end
-end
