--- app/sidekiq/task_stage/task_stage_overridden_job.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module TaskStage
-  class TaskStageOverriddenJob
-    include Sidekiq::Job
-
-    def perform(*args); end
-  end
-end
