+++ app/sidekiq/task_stage_errored_job.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class TaskStageErroredJob
+  include Sidekiq::Job
+
+  def perform(*args); end
+end
