+++ app/sidekiq/task_stage_failed_job.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class TaskStageFailedJob
+  include Sidekiq::Job
+
+  def perform(*_args)
+    Rails.logger.debug 'Ope, our task stage failed'
+  end
+end
