+++ app/sidekiq/task_stage_overridden_job.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class TaskStageOverriddenJob
+  include Sidekiq::Job
+
+  def perform(*args); end
+end
